Make things more generic

+++ build-tex.sh
@@ -1,3 +1,7 @@
 python3 generate_tex.py
 lualatex huy-cv.tex
-kill -SIGHUP "$(pgrep mupdf)"
+if [ -z "$(pgrep mupdf)" ]; then
+  mupdf huy-cv.pdf
+else
+  kill -SIGHUP "$(pgrep mupdf)"
+fi
